Fix build errors.

+++ src/mono/mono/eventpipe/ep-json-file.c
@@ -134,7 +134,7 @@ ep_json_file_write_event_data (
 			json_file_write_string (json_file, buffer);
 	}
 
-	characters_written = ep_rt_utf8_string_snprintf (buffer, EP_ARRAY_SIZE (buffer), "\"Thread (%ld)\"]},", thread_id);
+	characters_written = ep_rt_utf8_string_snprintf (buffer, EP_ARRAY_SIZE (buffer), "\"Thread (%ld)\"]},", (uint64_t)thread_id);
 	if (characters_written > 0 && characters_written < EP_ARRAY_SIZE (buffer))
 		json_file_write_string (json_file, buffer);
 }
